+++ benchmarks/build.gradle
@@ -36,7 +36,8 @@ dependencies {
             project(':grpc-netty'),
             project(':grpc-okhttp'),
             project(':grpc-stub'),
-            project(':grpc-interop-testing'),
+            project(':grpc-protobuf'),
+            project(':grpc-testing'),
             libraries.junit,
             libraries.mockito,
             libraries.hdrhistogram,
