+++ interop-testing/src/main/java/io/grpc/testing/integration/Util.java
@@ -38,23 +38,7 @@ import io.grpc.protobuf.ProtoUtils;
 
 import org.junit.Assert;
 
-import java.io.BufferedInputStream;
-import java.io.BufferedWriter;
-import java.io.File;
-import java.io.FileInputStream;
-import java.io.FileWriter;
-import java.io.IOException;
-import java.io.InputStream;
-import java.net.ServerSocket;
-import java.security.KeyStore;
-import java.security.cert.CertificateFactory;
-import java.security.cert.X509Certificate;
 import java.util.List;
-
-import javax.net.ssl.SSLContext;
-import javax.net.ssl.SSLSocketFactory;
-import javax.net.ssl.TrustManagerFactory;
-import javax.security.auth.x500.X500Principal;
 
 /**
  * Utility methods to support integration testing.
@@ -63,43 +47,6 @@ public class Util {
 
   public static final Metadata.Key<Messages.SimpleContext> METADATA_KEY =
       ProtoUtils.keyForProto(Messages.SimpleContext.getDefaultInstance());
-
-  /**
-   * Picks an unused port.
-   */
-  public static int pickUnusedPort() {
-    try {
-      ServerSocket serverSocket = new ServerSocket(0);
-      int port = serverSocket.getLocalPort();
-      serverSocket.close();
-      return port;
-    } catch (IOException e) {
-      throw new RuntimeException(e);
-    }
-  }
-
-  /**
-   * Load a file from the resources folder.
-   *
-   * @param name  name of a file in src/main/resources/certs.
-   */
-  public static File loadCert(String name) throws IOException {
-    InputStream in = Util.class.getResourceAsStream("/certs/" + name);
-    File tmpFile = File.createTempFile(name, "");
-    tmpFile.deleteOnExit();
-
-    BufferedWriter writer = new BufferedWriter(new FileWriter(tmpFile));
-    try {
-      int b;
-      while ((b = in.read()) != -1) {
-        writer.write(b);
-      }
-    } finally {
-      writer.close();
-    }
-
-    return tmpFile;
-  }
 
   /** Assert that two messages are equal, producing a useful message if not. */
   public static void assertEquals(MessageLite expected, MessageLite actual) {
@@ -128,27 +75,5 @@ public class Util {
         assertEquals(expected.get(i), actual.get(i));
       }
     }
-  }
-
-  /**
-   * Returns a SSLSocketFactory which uses the certificate specified in certChainFile.
-   */
-  public static SSLSocketFactory getSslSocketFactoryForCertainCert(File certChainFile)
-      throws Exception {
-    KeyStore ks = KeyStore.getInstance("JKS");
-    ks.load(null, null);
-    CertificateFactory cf = CertificateFactory.getInstance("X.509");
-    X509Certificate cert = (X509Certificate) cf.generateCertificate(
-        new BufferedInputStream(new FileInputStream(certChainFile)));
-    X500Principal principal = cert.getSubjectX500Principal();
-    ks.setCertificateEntry(principal.getName("RFC2253"), cert);
-
-    // Set up trust manager factory to use our key store.
-    TrustManagerFactory trustManagerFactory =
-        TrustManagerFactory.getInstance(TrustManagerFactory.getDefaultAlgorithm());
-    trustManagerFactory.init(ks);
-    SSLContext context = SSLContext.getInstance("TLS");
-    context.init(null, trustManagerFactory.getTrustManagers(), null);
-    return context.getSocketFactory();
   }
 }
